+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: snaomi <snaomi@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/07 07:16:47 by snaomi            #+#    #+#             */
-/*   Updated: 2020/07/25 09:15:19 by snaomi           ###   ########.fr       */
+/*   Updated: 2020/07/29 02:45:57 by snaomi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,7 @@ char				*ft_substr(char const *s, unsigned int start, size_t len);
 char				*ft_strjoin(char const *s1, char const *s2);
 char				*ft_strtrim(char const *s1, char const *set);
 char				**ft_split(char const *s, char c);
-char				*ft_xtoa(unsigned int n, char *base);
+char				*ft_xtoa(unsigned long long n, char *base);
 char				*ft_utoa(unsigned int n);	
 char				*ft_itoa(int n);
 char				*ft_strmapi(char const *s, char (*f)(unsigned int, char));
